+++ src/ui/input_forms.py
@@ -1,7 +1,7 @@
 import asyncio
+import logging
 
-from PyQt5.QtCore import QTimer
-from PyQt5.QtWidgets import (
+from PySide6.QtWidgets import (
     QComboBox,
     QFileDialog,
     QLabel,
@@ -11,6 +11,7 @@ from PyQt5.QtWidgets import (
     QVBoxLayout,
     QWidget,
 )
+from qasync import QEventLoop
 
 from core.emissions_calculator import EmissionsCalculator
 from data.data_validator import DataValidator
@@ -18,12 +19,13 @@ from data.database import get_fuel_types
 from data.export_manager import ExportManager
 from data.import_manager import ImportManager
 
+logger = logging.getLogger("ui")
+
 
 # this inherits from QWidget and DataValidator (this is to used to validate the data from user)
 class InputForms(QWidget):
     def __init__(self, parent=None):
         super().__init__(parent)
-
         self.user_id_label = QLabel("User ID:")
         self.user_id_entry = QLineEdit()
 
@@ -48,34 +50,21 @@ class InputForms(QWidget):
 
         self.setLayout(layout)
 
+        loop = QEventLoop()
+        asyncio.set_event_loop(loop)
+        asyncio.ensure_future(self.populate_fuel_types())
+
+    async def populate_fuel_types(self):
         try:
-            self.loop = asyncio.get_running_loop()
-        except RuntimeError:
-            self.loop = asyncio.new_event_loop()
-            asyncio.set_event_loop(self.loop)
-
-        self.loop.run_until_complete(self.fetch_fuel_data())
-        self.loop.close()
-
-        # Use QTimer to periodically check for the completion of the coroutine
-        self.timer = QTimer()
-        self.timer.timeout.connect(self.check_fuel_types_loaded)
-        self.timer.start(100)  # Check every 100 ms
-
-    async def fetch_fuel_data(self):
-        print("fetching")
-        try:
-            self.fuel_types = await get_fuel_types()
-            print(f"Fetched fuel types: {self.fuel_types}")
+            logger.info("Populating fuel types")
+            fuel_types = await get_fuel_types()
+            self.fuel_type_entry.addItems(fuel_types)
+            logger.info("Fuel types populated")
         except Exception as e:
-            print(f"Error fetching fuel types: {e}")
-            self.fuel_types = []
-
-    def check_fuel_types_loaded(self):
-        if hasattr(self, "fuel_types"):
-            print("found attr")
-            self.fuel_type_entry.addItems(self.fuel_types)
-            self.timer.stop()
+            logger.error(f"Error populating fuel types: {e}")
+            QMessageBox.critical(
+                self, "Error", str(f"{e}: Error populating fuel types")
+            )
 
     def submit(self):
         """
@@ -88,12 +77,21 @@ class InputForms(QWidget):
         Raises:
             QMessageBox: If any input is invalid, an error message is displayed.
         """
+        logger.info("Submit button clicked")
         user_id = self.user_id_entry.text()
         fuel_type = self.fuel_type_entry.currentText()
         fuel_used = self.fuel_used_entry.text()
-        user_id = int(user_id) if user_id else None
-        fuel_used = float(fuel_used) if fuel_used else None
         data_validator = DataValidator()
+        try:
+            user_id = int(user_id)
+        except ValueError:
+            logger.error("Invalid user ID")
+            user_id = None
+        try:
+            fuel_used = float(fuel_used)
+        except ValueError:
+            logger.error("Invalid fuel used value")
+            fuel_used = None
         try:
             # user_id, fuel_type and fuel_used are NOT valid, it will raise a ValueError and QMessageBox will show the error message
             if (
@@ -110,12 +108,13 @@ class InputForms(QWidget):
                 and data_validator.validate_fuel_type(fuel_type)
                 and data_validator.validate_fuel_used(fuel_used)
             ):
+                logger.info("Valid input: sending data to EmissionsCalculator")
                 # Send the validated data to EmissionsCalculator
                 emissions_calculator = EmissionsCalculator()
                 emissions = emissions_calculator.calculate_emissions(
                     user_id, fuel_type, fuel_used
                 )
-                self.results = f"User ID: {user_id}, Fuel Type: {fuel_type}, Fuel Used: {fuel_used}, Emissions: {emissions[3]}"
+                self.results = f"User ID: {user_id}, Fuel Type: {fuel_type}, Fuel Used: {fuel_used}, Emissions: {emissions}"
                 QMessageBox.information(self, "Info", self.results)
                 self.import_export_data()
         except ValueError as e:
@@ -143,13 +142,13 @@ class InputForms(QWidget):
                 "CSV Files (*.csv);;JSON Files (*.json);;All Files (*)",
             )
             if file_path:
-                print(f"Importing data from: {file_path}")
                 import_manager = ImportManager(file_path)
                 if file_path.endswith(".csv"):
                     import_manager.import_from_csv()
                 elif file_path.endswith(".json"):
                     import_manager.import_from_json()
                 else:
+                    logger.error("Unsupported file format")
                     QMessageBox.critical(self, "Error", "Unsupported file format")
         elif msg_box.clickedButton() == export_button:
             # Open file dialog to export data
@@ -160,11 +159,11 @@ class InputForms(QWidget):
                 "CSV Files (*.csv);;JSON Files (*.json);;All Files (*)",
             )
             if file_path:
-                print(f"Exporting data to: {file_path}")
                 export_manager = ExportManager("databases/emissions.db")
                 if file_path.endswith(".csv"):
                     export_manager.export_to_csv(file_path)
                 elif file_path.endswith(".json"):
                     export_manager.export_to_json(file_path)
                 else:
+                    logger.error("Unsupported file format")
                     QMessageBox.critical(self, "Error", "Unsupported file format")